Added the .ida to the gitignore. Renaned the navbar.js to Navbar.js. Moved the about page to the left.
